Resize frames on edge-device, use faster frame change detection, introduce max fps

+++ edge-server/model.py
@@ -20,7 +20,7 @@ class RawDetection:
 class Detection(BaseModel):
     category: str
     score: int
-    bbox: List[int]
+    bbox: List[float]
 
 
 class DetectionResponse(BaseModel):
